Split out the study activity event stuff from existing tests for the global event system.
