+++ src/test/java/org/sagebionetworks/bridge/sdk/integration/ActivityEventTest.java
@@ -1,18 +1,17 @@
 package org.sagebionetworks.bridge.sdk.integration;
 
-import static org.joda.time.DateTimeZone.UTC;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.sagebionetworks.bridge.rest.model.ActivityEventUpdateType.FUTURE_ONLY;
+import static org.sagebionetworks.bridge.rest.model.ActivityEventUpdateType.IMMUTABLE;
+import static org.sagebionetworks.bridge.rest.model.ActivityEventUpdateType.MUTABLE;
 import static org.sagebionetworks.bridge.sdk.integration.InitListener.EVENT_KEY1;
 import static org.sagebionetworks.bridge.sdk.integration.InitListener.EVENT_KEY2;
 import static org.sagebionetworks.bridge.sdk.integration.InitListener.EVENT_KEY3;
-import static org.sagebionetworks.bridge.sdk.integration.Tests.STUDY_ID_1;
-import static org.sagebionetworks.bridge.sdk.integration.Tests.STUDY_ID_2;
 
 import java.io.IOException;
 import java.util.List;
@@ -20,30 +19,28 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import com.google.common.collect.ImmutableSet;
-
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
+import org.joda.time.Period;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
 import org.sagebionetworks.bridge.rest.api.ForConsentedUsersApi;
+import org.sagebionetworks.bridge.rest.api.ForDevelopersApi;
 import org.sagebionetworks.bridge.rest.api.ForResearchersApi;
-import org.sagebionetworks.bridge.rest.exceptions.EntityNotFoundException;
 import org.sagebionetworks.bridge.rest.model.ActivityEvent;
 import org.sagebionetworks.bridge.rest.model.ActivityEventList;
+import org.sagebionetworks.bridge.rest.model.App;
 import org.sagebionetworks.bridge.rest.model.CustomActivityEventRequest;
 import org.sagebionetworks.bridge.rest.model.Role;
-import org.sagebionetworks.bridge.rest.model.StudyActivityEvent;
-import org.sagebionetworks.bridge.rest.model.StudyActivityEventList;
-import org.sagebionetworks.bridge.rest.model.StudyActivityEventPagedList;
-import org.sagebionetworks.bridge.rest.model.StudyActivityEventRequest;
 import org.sagebionetworks.bridge.rest.model.StudyParticipant;
 import org.sagebionetworks.bridge.user.TestUserHelper;
 
 public class ActivityEventTest {
-
+    private static final String TWO_WEEKS_AFTER_KEY = "2-weeks-after";
+    private static final String TWO_WEEKS_AFTER_VALUE = "enrollment:P2W";
+    
     private static TestUserHelper.TestUser developer;
     private static TestUserHelper.TestUser researcher;
     private static TestUserHelper.TestUser user;
@@ -56,7 +53,38 @@ public class ActivityEventTest {
         researchersApi = researcher.getClient(ForResearchersApi.class);
 
         developer = TestUserHelper.createAndSignInUser(ActivityEventTest.class, false, Role.DEVELOPER);
+        ForDevelopersApi developersApi = developer.getClient(ForDevelopersApi.class);
 
+        App app = developersApi.getUsersApp().execute().body();
+        boolean updateApp = false;
+
+        // Add custom event keys, if not already present, with three different update behaviors.
+        if (!app.getCustomEvents().keySet().contains(EVENT_KEY1) || 
+                app.getCustomEvents().get(EVENT_KEY1) != MUTABLE) {
+            app.getCustomEvents().put(EVENT_KEY1, MUTABLE);
+            updateApp = true;
+        }
+        if (!app.getCustomEvents().keySet().contains(EVENT_KEY2) ||
+                app.getCustomEvents().get(EVENT_KEY2) != IMMUTABLE) {
+            app.getCustomEvents().put(EVENT_KEY2, IMMUTABLE);
+            updateApp = true;
+        }
+        if (!app.getCustomEvents().keySet().contains(EVENT_KEY3) ||
+                app.getCustomEvents().get(EVENT_KEY3) != FUTURE_ONLY) {
+            app.getCustomEvents().put(EVENT_KEY3, FUTURE_ONLY);
+            updateApp = true;
+        }
+
+        // Add automatic custom event.
+        if (!app.getAutomaticCustomEvents().containsKey(TWO_WEEKS_AFTER_KEY)) {
+            app.putAutomaticCustomEventsItem(TWO_WEEKS_AFTER_KEY, TWO_WEEKS_AFTER_VALUE);
+            updateApp = true;
+        }
+
+        if (updateApp) {
+            developersApi.updateUsersApp(app).execute();
+        }
+        
         // Create user last, so the automatic custom events are created
         user = TestUserHelper.createAndSignInUser(ActivityEventTest.class, true);
         usersApi = user.getClient(ForConsentedUsersApi.class);
@@ -151,28 +179,28 @@ public class ActivityEventTest {
                 .collect(Collectors.toSet()).contains(EVENT_KEY1));
     }
 
-//    @Test
-//    public void automaticCustomEvents() throws Exception {
-//        // Get activity events and convert to map for ease of use
-//        List<ActivityEvent> activityEventList = usersApi.getActivityEvents().execute().body().getItems();
-//        Map<String, ActivityEvent> activityEventMap = activityEventList.stream().collect(
-//                Collectors.toMap(ActivityEvent::getEventId, e -> e));
-//
-//        // Verify enrollment events exist
-//        ActivityEvent enrollmentEvent = activityEventMap.get("enrollment");
-//        assertNotNull(enrollmentEvent);
-//        DateTime enrollmentTime = enrollmentEvent.getTimestamp();
-//        assertNotNull(enrollmentTime);
-//
-//        // Verify custom event exists and that it's 2 weeks after enrollment
-//        ActivityEvent twoWeeksAfterEvent = activityEventMap.get("custom:" + TWO_WEEKS_AFTER_KEY);
-//        assertNotNull(twoWeeksAfterEvent);
-//        DateTime twoWeeksAfterTime = twoWeeksAfterEvent.getTimestamp();
-//        // This can fail when you're near the time zone change to DST. Add one hour to overshoot 
-//        // and compensate for the time zone change.
-//        Period twoWeeksAfterPeriod = new Period(enrollmentTime, twoWeeksAfterTime.plusHours(1));
-//        assertEquals(2, twoWeeksAfterPeriod.getWeeks());
-//    }
+    @Test
+    public void automaticCustomEvents() throws Exception {
+        // Get activity events and convert to map for ease of use
+        List<ActivityEvent> activityEventList = usersApi.getActivityEvents().execute().body().getItems();
+        Map<String, ActivityEvent> activityEventMap = activityEventList.stream().collect(
+                Collectors.toMap(ActivityEvent::getEventId, e -> e));
+
+        // Verify enrollment events exist
+        ActivityEvent enrollmentEvent = activityEventMap.get("enrollment");
+        assertNotNull(enrollmentEvent);
+        DateTime enrollmentTime = enrollmentEvent.getTimestamp();
+        assertNotNull(enrollmentTime);
+
+        // Verify custom event exists and that it's 2 weeks after enrollment
+        ActivityEvent twoWeeksAfterEvent = activityEventMap.get("custom:" + TWO_WEEKS_AFTER_KEY);
+        assertNotNull(twoWeeksAfterEvent);
+        DateTime twoWeeksAfterTime = twoWeeksAfterEvent.getTimestamp();
+        // This can fail when you're near the time zone change to DST. Add one hour to overshoot 
+        // and compensate for the time zone change.
+        Period twoWeeksAfterPeriod = new Period(enrollmentTime, twoWeeksAfterTime.plusHours(1));
+        assertEquals(2, twoWeeksAfterPeriod.getWeeks());
+    }
     
     @Test
     public void researcherCanSubmitCustomEvents() throws Exception {
@@ -202,85 +230,6 @@ public class ActivityEventTest {
                 .filter((evt) -> evt.getEventId().equals("custom:"+EVENT_KEY1)).findAny();
         assertTrue(optional.isPresent());
         assertEquals(timestamp2.toString(), optional.get().getTimestamp().toString());
-    }
-    
-    @Test
-    public void createStudyActivityEvent() throws Exception {
-        // Because this is set first, it is immutable and will not be changed by the study-scoped value.
-        // Both with exist side-by-side
-        DateTime globalTimestamp = DateTime.now(UTC).minusDays(2);
-        CustomActivityEventRequest globalRequest = new CustomActivityEventRequest()
-                .eventId(EVENT_KEY1).timestamp(globalTimestamp);
-        researchersApi.createActivityEventForParticipant(user.getUserId(), globalRequest).execute();
-        
-        DateTime studyScopedTimestamp = DateTime.now(UTC);  
-        StudyActivityEventRequest studyScopedRequest = new StudyActivityEventRequest()
-            .eventId(EVENT_KEY1).timestamp(studyScopedTimestamp);
-        researchersApi.createStudyParticipantStudyActivityEvent(STUDY_ID_1, user.getUserId(), studyScopedRequest).execute();
-        
-        ActivityEventList globalList = researchersApi.getActivityEventsForParticipant(user.getUserId()).execute().body();
-        ActivityEvent globalEvent = findEventByKey(globalList, "custom:"+EVENT_KEY1);
-        assertEquals(globalTimestamp, globalEvent.getTimestamp());
-        
-        StudyActivityEventList scopedList = researchersApi.getStudyParticipantStudyActivityEvents(STUDY_ID_1, user.getUserId()).execute().body();
-        StudyActivityEvent scopedEvent = findEventByKey(scopedList, "custom:"+EVENT_KEY1);
-        assertEquals(studyScopedTimestamp, scopedEvent.getTimestamp());
-
-        // User can also retrieve these two different events.
-        scopedList = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        scopedEvent = findEventByKey(scopedList, "custom:"+EVENT_KEY1);
-        assertEquals(studyScopedTimestamp, scopedEvent.getTimestamp());
-        
-        globalList= usersApi.getActivityEvents().execute().body();
-        globalEvent = findEventByKey(globalList, "custom:"+EVENT_KEY1);
-        assertEquals(globalTimestamp, globalEvent.getTimestamp());
-        
-        // Verify that a user can create a custom event, and that it is visible to researchers
-        globalTimestamp = globalTimestamp.minusWeeks(2);
-        globalRequest = new CustomActivityEventRequest().eventId(EVENT_KEY2).timestamp(globalTimestamp);
-        usersApi.createCustomActivityEvent(globalRequest).execute();
-        
-        studyScopedTimestamp = studyScopedTimestamp.minusWeeks(2);
-        studyScopedRequest = new StudyActivityEventRequest().eventId(EVENT_KEY2).timestamp(studyScopedTimestamp);
-        usersApi.createStudyActivityEvent(STUDY_ID_1, studyScopedRequest).execute();
-
-        // user can see these events
-        scopedList = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        scopedEvent = findEventByKey(scopedList, "custom:"+EVENT_KEY2);
-        assertEquals(studyScopedTimestamp, scopedEvent.getTimestamp());
-        
-        globalList= usersApi.getActivityEvents().execute().body();
-        globalEvent = findEventByKey(globalList, "custom:"+EVENT_KEY2);
-        assertEquals(globalTimestamp, globalEvent.getTimestamp());
-
-        // the researcher can see these events
-        globalList = researchersApi.getActivityEventsForParticipant(user.getUserId()).execute().body();
-        globalEvent = findEventByKey(globalList, "custom:"+EVENT_KEY2);
-        assertEquals(globalTimestamp, globalEvent.getTimestamp());
-        
-        scopedList = researchersApi.getStudyParticipantStudyActivityEvents(STUDY_ID_1, user.getUserId()).execute().body();
-        scopedEvent = findEventByKey(scopedList, "custom:"+EVENT_KEY2);
-        assertEquals(studyScopedTimestamp, scopedEvent.getTimestamp());
-    }
-    
-    @Test
-    public void globalEventsDoNotCreateStudyVersions() throws Exception {
-        DateTime globalTimestamp = DateTime.now(UTC).minusDays(2);
-        CustomActivityEventRequest globalRequest = new CustomActivityEventRequest()
-                .eventId(EVENT_KEY1).timestamp(globalTimestamp);
-        researchersApi.createActivityEventForParticipant(user.getUserId(), globalRequest).execute();
-
-        StudyActivityEventList scopedList = researchersApi.getStudyParticipantStudyActivityEvents(
-                STUDY_ID_1, user.getUserId()).execute().body();
-        assertNull(findEventByKey(scopedList, "custom:"+EVENT_KEY1));
-
-        // The user has not been enrolled in this study so there should be no events.
-        try {
-            researchersApi.getStudyParticipantStudyActivityEvents(STUDY_ID_2, user.getUserId()).execute().body();
-            fail("Should have thrown an exception.");
-        } catch(EntityNotFoundException e) {
-            
-        }
     }
     
     @Test
@@ -378,134 +327,6 @@ public class ActivityEventTest {
         assertEquals(getTimestamp(list3, EVENT_KEY3), futureTime);
     }
     
-    @Test
-    public void testMutableStudyEvent() throws Exception {
-        DateTime now = DateTime.now(DateTimeZone.UTC);
-        DateTime pastTime = DateTime.now(DateTimeZone.UTC).minusHours(2);
-        DateTime futureTime = DateTime.now(DateTimeZone.UTC).plusHours(2);
-
-        // Create event #1 which is mutable
-        StudyActivityEventRequest req = new StudyActivityEventRequest().eventId(EVENT_KEY1).timestamp(now);
-        usersApi.createStudyActivityEvent(STUDY_ID_1, req).execute();
-        
-        StudyActivityEventList list = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(now, getTimestamp(list, EVENT_KEY1));
-        
-        // future time updates
-        req = new StudyActivityEventRequest().eventId(EVENT_KEY1).timestamp(futureTime);
-        usersApi.createStudyActivityEvent(STUDY_ID_1, req).execute();
-        list = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(futureTime, getTimestamp(list, EVENT_KEY1));
-
-        // past time updates
-        req = new StudyActivityEventRequest().eventId(EVENT_KEY1).timestamp(pastTime);
-        usersApi.createStudyActivityEvent(STUDY_ID_1, req).execute();
-        list = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(pastTime, getTimestamp(list, EVENT_KEY1));
-        
-        // This will have accumulated three events
-        StudyActivityEventPagedList page = usersApi.getStudyActivityEventHistory(
-                STUDY_ID_1, EVENT_KEY1, null, null).execute().body();
-        assertEquals(3, page.getItems().size());
-        assertEquals(Integer.valueOf(3), page.getTotal());
-        assertEquals(ImmutableSet.of(now, pastTime, futureTime), page.getItems().stream()
-                .map(StudyActivityEvent::getTimestamp).collect(Collectors.toSet()));
-        
-        // These are counted in the record total
-        list = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(Integer.valueOf(3), findEventByKey(list, "custom:"+EVENT_KEY1).getRecordCount());
-
-        // can delete
-        usersApi.deleteStudyActivityEvent(STUDY_ID_1, EVENT_KEY1).execute();
-        
-        list = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertNull(getTimestamp(list, EVENT_KEY1));
-        
-        // Cannot retrieve events for study 2.
-        try {
-            usersApi.getStudyActivityEvents(STUDY_ID_2).execute().body();
-            fail("Should have thrown exception");
-        } catch(EntityNotFoundException e) {
-            assertEquals("Account not found.", e.getMessage());
-        }
-    }
-
-    @Test
-    public void testImmutableStudyEvent() throws Exception {
-        DateTime now = DateTime.now(DateTimeZone.UTC);
-        DateTime pastTime = DateTime.now(DateTimeZone.UTC).minusHours(2);
-        DateTime futureTime = DateTime.now(DateTimeZone.UTC).plusHours(2);
-
-        // Create event #2 which is immutable.
-        StudyActivityEventRequest req = new StudyActivityEventRequest().eventId(EVENT_KEY2).timestamp(now);
-        usersApi.createStudyActivityEvent(STUDY_ID_1, req).execute();
-        StudyActivityEventList list = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(now, getTimestamp(list, EVENT_KEY2));
-
-        // past time won't update
-        req = new StudyActivityEventRequest().eventId(EVENT_KEY2).timestamp(pastTime);
-        usersApi.createStudyActivityEvent(STUDY_ID_1, req).execute();
-        list = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(getTimestamp(list, EVENT_KEY2), now);
-
-        // future time won't update
-        req = new StudyActivityEventRequest().eventId(EVENT_KEY2).timestamp(futureTime);
-        usersApi.createStudyActivityEvent(STUDY_ID_1, req).execute();
-        list = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(getTimestamp(list, EVENT_KEY2), now);
-
-        // nor will it delete
-        usersApi.deleteStudyActivityEvent(STUDY_ID_1, EVENT_KEY2).execute();
-
-        list = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(getTimestamp(list, EVENT_KEY2), now);
-        
-        // nor will it accumulate in history
-        StudyActivityEventPagedList page = usersApi.getStudyActivityEventHistory(
-                STUDY_ID_1, EVENT_KEY2, null, null).execute().body();
-        assertEquals(1, page.getItems().size());
-        assertEquals(Integer.valueOf(1), page.getTotal());
-    }
-    
-    @Test
-    public void testFutureOnlyStudyEvent() throws Exception {
-        DateTime now = DateTime.now(DateTimeZone.UTC);
-        DateTime pastTime = DateTime.now(DateTimeZone.UTC).minusHours(2);
-        DateTime futureTime = DateTime.now(DateTimeZone.UTC).plusHours(2);
-
-        StudyActivityEventRequest req3 = new StudyActivityEventRequest().eventId(EVENT_KEY3).timestamp(now);
-        usersApi.createStudyActivityEvent(STUDY_ID_1, req3).execute();
-        
-        StudyActivityEventList list3 = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(now, getTimestamp(list3, EVENT_KEY3));
-        
-        // future will update
-        req3 = new StudyActivityEventRequest().eventId(EVENT_KEY3).timestamp(futureTime);
-        usersApi.createStudyActivityEvent(STUDY_ID_1, req3).execute();
-        list3 = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(getTimestamp(list3, EVENT_KEY3), futureTime);
-        
-        // past will not update
-        req3 = new StudyActivityEventRequest().eventId(EVENT_KEY3).timestamp(pastTime);
-        usersApi.createStudyActivityEvent(STUDY_ID_1, req3).execute();
-        list3 = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(getTimestamp(list3, EVENT_KEY3), futureTime);
-
-        // This doesn't delete the timestamp
-        usersApi.deleteStudyActivityEvent(STUDY_ID_1, EVENT_KEY3).execute();
-        
-        list3 = usersApi.getStudyActivityEvents(STUDY_ID_1).execute().body();
-        assertEquals(getTimestamp(list3, EVENT_KEY3), futureTime);
-        
-        // This will have accumulated two events
-        StudyActivityEventPagedList page = usersApi.getStudyActivityEventHistory(
-                STUDY_ID_1, EVENT_KEY3, null, null).execute().body();
-        assertEquals(2, page.getItems().size());
-        assertEquals(Integer.valueOf(2), page.getTotal());
-        assertEquals(ImmutableSet.of(now, futureTime), page.getItems().stream()
-                .map(StudyActivityEvent::getTimestamp).collect(Collectors.toSet()));
-    }
-    
     private DateTime getTimestamp(ActivityEventList list, String eventId) {
         for (ActivityEvent event : list.getItems()) {
             if (event.getEventId().equals("custom:"+eventId)) {
@@ -514,31 +335,5 @@ public class ActivityEventTest {
         }
         return null;
     }
-    
-    private DateTime getTimestamp(StudyActivityEventList list, String eventId) {
-        for (StudyActivityEvent event : list.getItems()) {
-            if (event.getEventId().equals("custom:"+eventId)) {
-                return event.getTimestamp();
-            }
-        }
-        return null;
-    }
-    
-    private ActivityEvent findEventByKey(ActivityEventList list, String key) {
-        for (ActivityEvent event : list.getItems()) {
-            if (event.getEventId().equals(key)) {
-                return event;
-            }
-        }
-        return null;
-    }
-    
-    private StudyActivityEvent findEventByKey(StudyActivityEventList list, String key) {
-        for (StudyActivityEvent event : list.getItems()) {
-            if (event.getEventId().equals(key)) {
-                return event;
-            }
-        }
-        return null;
-    }
+
 }
